+++ Editor/ArenaVersion.cs
@@ -1,0 +1,152 @@
+/**
+ * Open source software under the terms in /LICENSE
+ * Copyright (c) 2021, The CONIX Research Center. All rights reserved.
+ */
+
+using System;
+using System.Collections;
+using System.Linq;
+using Newtonsoft.Json;
+using Unity.EditorCoroutines.Editor;
+using UnityEditor;
+using UnityEditor.PackageManager;
+using UnityEditor.PackageManager.Requests;
+using UnityEngine;
+using UnityEngine.Networking;
+
+namespace ArenaUnity.Editor
+{
+    /// <summary>
+    /// Editor class to check installed package versions against new releases.
+    /// </summary>
+    [InitializeOnLoad]
+    internal class ArenaVersion
+    {
+        const string unityPackageName = "io.conix.arena.unity";
+        const string githubOrg = "conix-center";
+        const string githubName = "ARENA-unity";
+        private static string gitLatestUrl = $"https://api.github.com/repos/{githubOrg}/{githubName}/releases/latest";
+        private static ListRequest _listRequest;
+        private static bool checkGihub = false;
+
+        // first version to avoid hitting GiHub's rate-limit
+        private const string GH_RATE_LIMIT_VERSION = "0.0.4";
+
+        static ArenaVersion()
+        {
+            string latest = PlayerPrefs.GetString("GitVersionLatest", GH_RATE_LIMIT_VERSION);
+            long time = (long)PlayerPrefs.GetFloat("GitVersionCheckTime", 0);
+            TimeSpan t = DateTime.UtcNow - new DateTime(time);
+            // only check github every 24 hours to avoid hitting api rate limit
+            if (t.TotalDays > 1f) checkGihub = true;
+
+            _listRequest = Client.List();
+            EditorApplication.update += OnUpdate;
+        }
+
+        private static void HandleListRequestCompletion()
+        {
+            if (_listRequest.Status == StatusCode.Success)
+            {
+                var package = _listRequest.Result.FirstOrDefault(p => p.name == unityPackageName);
+                if (package != null && Version.TryParse(package.version, out var local))
+                {
+                    if (Version.TryParse(package.versions.latest, out var latest))
+                    {
+                        if (local < latest)
+                        {
+                            Debug.LogWarning(UpgradeMessage(local, latest));
+                        }
+                    }
+                    // Check github directly next
+                    if (checkGihub)
+                    {
+                        EditorCoroutineUtility.StartCoroutineOwnerless(CheckGithubVersion(local));
+                    }
+                    Debug.Log(CurrentMessage(local));
+                }
+                _listRequest = null;
+            }
+        }
+
+        static IEnumerator CheckGithubVersion(Version local)
+        {
+            UnityWebRequest www = UnityWebRequest.Get(gitLatestUrl);
+            yield return www.SendWebRequest();
+
+#if UNITY_2020_1_OR_NEWER
+            if (www.result == UnityWebRequest.Result.ConnectionError || www.result == UnityWebRequest.Result.ProtocolError)
+#else
+            if (www.isNetworkError || www.isHttpError)
+#endif
+            {
+                Debug.Log(www.error);
+            }
+            else
+            {
+                dynamic git = JsonConvert.DeserializeObject(www.downloadHandler.text);
+                if (git != null)
+                {
+                    if (Version.TryParse((string)git.tag_name, out var latest))
+                    {
+                        PlayerPrefs.SetString("GitVersionLatest", (string)git.tag_name);
+                        if (local < latest)
+                        {
+                            Debug.LogWarning(UpgradeMessage(local, latest));
+                        }
+                    }
+                }
+            }
+            PlayerPrefs.SetFloat("GitVersionCheckTime", DateTimeOffset.Now.Ticks);
+            PlayerPrefs.Save();
+        }
+
+        private static string UpgradeMessage(Version local, Version latest)
+        {
+            return $"ARENA for Unity Package version {latest} is available, however {local} is installed.\nUpdate to https://github.com/{githubOrg}/{githubName}.git#{latest}";
+        }
+
+        private static string CurrentMessage(Version local)
+        {
+            return $"ARENA for Unity Package version {local} is installed.\nLatest: https://github.com/{githubOrg}/{githubName}/releases";
+        }
+
+        private static void OnUpdate()
+        {
+            if (_listRequest != null && _listRequest.IsCompleted)
+            {
+                HandleListRequestCompletion();
+            }
+        }
+
+        internal static string LocalVersion()
+        {
+            var package = UnityEditor.PackageManager.PackageInfo.FindForAssembly(typeof(ArenaVersion).Assembly);
+            return package.version;
+        }
+
+        static bool WantsToQuit()
+        {
+            Debug.Log("Exiting play mode...");
+            EditorApplication.ExitPlaymode();
+
+            // Check that this instance was actually launched from a batch mode session, so that game code
+            // doesn't inadvertently exit the editor during development.
+            if (Application.isBatchMode)
+            {
+                EditorApplication.Exit(0);
+            }
+            else
+            {
+                Debug.Log("Exiting application.");
+            }
+            return true;
+        }
+
+        [RuntimeInitializeOnLoadMethod]
+        static void RunOnStart()
+        {
+            Application.wantsToQuit += WantsToQuit;
+        }
+    }
+}