+++ Runtime/ArenaObject.cs
@@ -1,15 +1,21 @@
-﻿using System;
+﻿/**
+ * Open source software under the terms in /LICENSE
+ * Copyright (c) 2021, The CONIX Research Center. All rights reserved.
+ */
+
+using System;
+using System.Dynamic;
 using Newtonsoft.Json;
 using UnityEngine;
-using System.Dynamic;
 
 namespace ArenaUnity
 {
+    /// <summary>
+    /// Class to manage an ARENA object, publishing, and its properties.
+    /// </summary>
     [HelpURL("https://arena.conix.io/content/messaging/definitions.html")]
     public class ArenaObject : MonoBehaviour
     {
-        [Tooltip("A uuid or otherwise unique identifier for this object")]
-        public string objectId = Guid.NewGuid().ToString();
         [Tooltip("Type in storage schema (RO)")]
         public string storeType = "entity"; // default to entity
         [Tooltip("Persist this object in the ARENA server database (default false = do not persist)")]
@@ -25,17 +31,14 @@ namespace ArenaUnity
         [HideInInspector]
         public bool created = false;
 
+        private string oldName; // test for rename
+
         public void OnEnable()
         {
-            if (ArenaClient.Instance == null || !ArenaClient.Instance.mqttClientConnected)
-                return;
-            // trigger publish for new object
-            transform.hasChanged = true;
         }
 
         void Start()
         {
-            transform.hasChanged = false;
         }
 
         void Update()
@@ -51,6 +54,29 @@ namespace ArenaUnity
                     transform.hasChanged = false;
                 }
             }
+            else if (oldName != null && name != oldName)
+            {
+                HandleRename();
+            }
+            oldName = name;
+        }
+
+        private void HandleRename()
+        {
+            if (ArenaClient.Instance == null || !ArenaClient.Instance.mqttClientConnected)
+                return;
+            // pub delete old
+            dynamic msg = new
+            {
+                object_id = oldName,
+                action = "delete",
+                persist = persist,
+            };
+            string payload = JsonConvert.SerializeObject(msg);
+            ArenaClient.Instance.Publish(msg.object_id, payload);
+            // add new object with new name, it pubs
+            created = false;
+            transform.hasChanged = true;
         }
 
         bool SendUpdateSuccess()
@@ -59,36 +85,33 @@ namespace ArenaUnity
                 return false;
 
             dynamic msg = new ExpandoObject();
-            msg.object_id = this.objectId;
-            msg.action = this.created ? "update" : "create";
-            msg.type = this.storeType;
-            msg.persist = this.persist;
+            msg.object_id = name;
+            msg.action = created ? "update" : "create";
+            msg.type = storeType;
+            msg.persist = persist;
 
             dynamic dataUp = new ExpandoObject();
-            if (data.object_type == null)
-                dataUp.object_type = ArenaUnity.ToArenaObjectType(this.gameObject);
+            if (data == null || data.object_type == null)
+                dataUp.object_type = ArenaUnity.ToArenaObjectType(gameObject);
+            else
+                dataUp.object_type = (string)data.object_type;
             dataUp.position = ArenaUnity.ToArenaPosition(transform.position);
-            if (data.rotation == null || data.rotation.w != null)
+            if (data == null || data.rotation == null || data.rotation.w != null)
                 dataUp.rotation = ArenaUnity.ToArenaRotationQuat(transform.rotation);
             else
                 dataUp.rotation = ArenaUnity.ToArenaRotationEuler(transform.rotation.eulerAngles);
-            dataUp.scale = ArenaUnity.ToArenaScale((string)data.object_type, transform.localScale);
+            dataUp.scale = ArenaUnity.ToArenaScale(transform.localScale);
+            ArenaUnity.ToArenaDimensions(gameObject, ref dataUp);
+            if (GetComponent<Light>())
+                ArenaUnity.ToArenaLight(gameObject, ref dataUp);
             if (GetComponent<Renderer>())
-            {
-                Color color = GetComponent<Renderer>().material.GetColor("_Color");
-                if (color != null)
-                {
-                    dynamic material = new ExpandoObject();
-                    material.color = ArenaUnity.ToArenaColor(color);
-                    dataUp.material = material;
-                }
-            }
+                ArenaUnity.ToArenaMaterial(gameObject, ref dataUp);
             msg.data = dataUp;
             //jsonData = JsonConvert.SerializeObject(data);
             string payload = JsonConvert.SerializeObject(msg);
             ArenaClient.Instance.Publish(msg.object_id, payload);
-            if (!this.created)
-                this.created = true;
+            if (!created)
+                created = true;
 
             return true;
         }
@@ -100,8 +123,9 @@ namespace ArenaUnity
 
             dynamic msg = new
             {
-                object_id = this.objectId,
+                object_id = name,
                 action = "delete",
+                persist = persist,
             };
             string payload = JsonConvert.SerializeObject(msg);
             ArenaClient.Instance.Publish(msg.object_id, payload);