+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 ARENA-unity notable changes. Started 2021-11-30 (version 0.0.1).
 
+## [0.0.4] - 2021-12-07
+### Changed
+- Reduced frequency of GitHub version checking.
+
+## [0.0.3] - 2021-12-06
+### Added
+- Version upgrade reminder direct from latest GitHub release tag.
+
+## [0.0.2] - 2021-12-06
+### Added
+- Allow `.glb` GLTF load.
+- ARENA cameras added to displays.
+- Consistent naming`GameObject.name` == `object_id`
+- Allow renaming of ARENA objects.
+- Option to prevent logging of non-persist objects.
+- Package version check in Editor.
+- Fixed Unity-side object publish.
+
 ## [0.0.1] - 2021-11-30
 ### Added
 - M2MqttUnity for ARENA TLS auth.